Implement playRound Functionality
Implemented the playRound function to simulate rounds being played by
the computer and a user. The function takes two args humanChoice and
computerChoice. I then create an if block for all scenarios of the game
and also increment global variables humanScore and computerScore by 1
depending on who won that round. Then allocate the other global variables
humanSelection and computerSelection to previous functions
getHumanChoice and getComputerChoice respectively. Then I call the
playRound function to test it.