Implemented the playGame Function
Implementing the playGame function so that multiple rounds can be played
between the user and computer. Nested the playRound function into the
playGame function so that it had access to the playRound variables. Also
moved the humanScore and computerScore from a global variable to a local
one inside of playGame. Then I called the playGame function to test
functionality.